aws-sigv4: enable the usage of s3 with curl
As describe here:
https://docs.aws.amazon.com/AmazonS3/latest/API/sig-v4-header-based-auth.html
It seems s3 require an additional x-amz-content-sha256 field for the signature.
So I modify the algorithm to check if 'X-Amz-Content-Sha256'
is present in the headers,
if so I use it to add all require information to the request.

Note: the amazone doc is only about s3, so I can't use
x-amz-content-sha256 in every request, as this would break the
signature for other APIs

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>