http-aws-sigv4: refactor headers
Up to this patch, canonical header, and signed header was hardcoded
in the aws-sigv4 signature.

This patch handle canonical headers and signed headers creation
as it is explain here:
https://docs.aws.amazon.com/general/latest/gr/sigv4-create-canonical-request.html

The algo tell that signed and canonical must contain at last host and x-amz-date.

So we check whatever thoses are present in the curl http headers list.
If they are, we use the one enter by curl user, otherwise we generate them.
then we to lower, and remove space from each http headers plus host and x-amz-date,
then sort them all by alphabetical order.

This patch also fix a bug with host header, which was ignoring the port.

Because I now handle the port in the signature,
I had to hardcode the host in the tests, in which the port was random,
which was making imposible to had a signature in the tests data.

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>